requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
langchain>=0.2
langchain-community>=0.2
langchain-ollama>=0.0.8
httpx>=0.27
anyio>=4.4
pydantic~=2.11.7
requests~=2.32.4
python-dotenv>=1.0